+++ CHANGELOG.md
@@ -2,4 +2,6 @@
 
 ## Unreleased
 
+- Initial (work in progress) version of GraphQL over gRPC support.
+
 - Token management & persistent storage